requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0                  
google-genai>=0.8.0              
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.8.0
python-multipart>=0.0.18                  